Simplify file extension comparison in checkjndi.ps1
Instead of chaining together `($_.Extension -eq '.foo') -or ($_.Extension -eq '.bar')` statements, you can use the `-in` comparison operator and provide an array of strings to check against: `$_.Extension -in ('.foo', '.bar')` This makes visualizing and editing the list easier and keeps the code concise, while still being readable.

Another advantage to this method is that, later, you could parameterize that array of extensions and pass it in as an argument to the script. Although, I won't include that functionality in this pull request.